+++ src/lib/pages/code-details/data.ts
@@ -3,7 +3,10 @@ import type { UseQueryOptions } from "@tanstack/react-query";
 import { useContractStore } from "lib/providers/store";
 import type { ContractsResponse } from "lib/services/types";
 import { useCodeLcd } from "lib/services/wasm/code";
-import { useContractsByCodeId } from "lib/services/wasm/contract";
+import {
+  useContractsByCodeId,
+  useContractsByCodeIdLcd,
+} from "lib/services/wasm/contract";
 import type { ContractInfo, Option } from "lib/types";
 
 export const useCodeDataLcd = (codeId: number, enabled: boolean) => {
@@ -50,5 +53,20 @@ export const useCodeContracts = (
       items: contracts,
     },
     isLoading,
+  };
+};
+
+export const useCodeContractsLcd = (codeId: number) => {
+  const { getContractLocalInfo } = useContractStore();
+  const { data, ...rest } = useContractsByCodeIdLcd(codeId);
+
+  return {
+    data: data?.pages.flatMap((page) =>
+      page.contracts.map<ContractInfo>((contract) => ({
+        ...contract,
+        ...getContractLocalInfo(contract.contractAddress),
+      }))
+    ),
+    ...rest,
   };
 };